server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pillow==11.3.0
numpy
tensorflow==2.19.0
keras==3.10.0
faiss-cpu==1.11.0.post1
langchain==0.3.27
langchain-community==0.3.27
langchain-mistralai==0.2.11
sentence-transformers==5.1.0
mistralai==1.9.3
huggingface-hub==0.34.4
pydantic==2.11.7
python-multipart==0.0.20
